+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Changed
 * Form Group: Errormessage error icon ([#2640](https://github.com/dso-toolkit/dso-toolkit/issues/2640))
 
+### Docs
+* Card: Slide Toggle toevoegen ([#2516](https://github.com/dso-toolkit/dso-toolkit/issues/2516))
+
 ### Tasks
 * Packages: Dependencies updates ([#2648](https://github.com/dso-toolkit/dso-toolkit/issues/2648))
 * Packages: `http-proxy-middleware` updaten naar ^3.0.0 ([#2612](https://github.com/dso-toolkit/dso-toolkit/issues/2612))